Fixed abort logic so that the code does not hang when a fatal error occurs.  Simplified ice_communicate.F90 (both serial and mpi) by removing the exit_message_environment and abort_message_environment routines, which were hold-overs from pop code imported into cice 4.0 that duplicated functionality already present in cice.
MPI barriers can not be used when aborting cice because of the
very many instances where one processor can abort without the 
others. Broadcasting the l_stop flag prior to aborting does not 
help because other processors may already be past that point 
in the code; they might then write faulty diagnostics (confusing 
the user) and/or hang elsewhere, waiting for the aborted processor.