Fixed hardwired 365-day year for the UK Metoffice, which uses a 360-day year.
   ice_calendar.F    Allow either 360 or 365 days per year 
   ice_init.F        New namelist option:  days_per_year (default=365)
   ice_history.F     Generalized netcdf comment
                     Changed iday=31 to iday=daymo(12)
   ice_constants.F   Removed unused conversion factor, mps_to_cmpyr
   ice_forcing.F     Allow interpolations to handle a 360-day year.  365 is
      still hardwired where appropriate for particular forcing data sets.
   ice_in            New namelist option:  days_per_year

The current implementation of the delta-Eddington (dEdd) shortwave requires
a 365-day year because it is based on CCSM shared code.

I am not confident that I found _all_ instances that assume a 365-day year.
The 365-day year option is BFB in my standard configuration.  The 360-day 
year option needs more testing.