Corrected indices in boundary_3d_dbl. Corrected 0 type in boundary_2d_int.
I think this fixes the boundary conditions---at least the problem I discovered this 
morning---so this revision should be okay to use.  The boundary condition routines
are fairly intricate, and I would not be surprised if there are still more bugs. If 
any of you notices anything suspicious, please let me know.  --Elizabeth