Checked in a new version of ice_transport_remap.F which allows the user to specify the geometric area flux across each cell edge.  This means that the divergence computed by the EVP dynamics and the divergence implied by the remapping can be made identical.  The method is based on an idea of Mats Bentsen.
For now, the older remapping routine from CICE 3.14 is still in the code.
To call the old routine, set newremap = F in ice_transport_driver.
(This is the default for now.)  After the new code has been tested
sufficiently, the old code will be removed.  The two versions give
the same answers within roundoff in simple advection tests on a rectangular
grid.  On the displaced-pole grid the answers differ slightly because
some of the geometry assumptions are different.

Other changes are minor and do not change the answers.