Merged ice_therm_vertical.F90 r327:328 from ucl branch into trunk.
These modifications to the thermodynamics handle the case in which enough 
heat collects in an ice layer to completely melt it.   This case normally 
resulted in a thermodynamic convergence error.   Now, if the available 
energy would melt through a layer, then penetrating shortwave is first 
reduced, possibly to zero, and if that is insufficient then the local 
conductivity is also reduced to bring the layer temperature just to the 
melting point.

These changes essentially relax the thermo convergence criteria, and they 
have been successful in runs using the ucl and zbgc branch versions of the 
code.  If any of you encounter problems with this approach, let me know 
pronto! eclare@lanl.gov

M source/ice_therm_vertical.F90
- increase nitermax to 100
- allow shortwave radiation to melt only 90% of any layer (was 1-puny)
- increase initial temperature difference from freezing, for limiting 
shortwave, to 0.02C
- include energy due to resetting temperature to the freezing point in 
the convergence test
- reduce conductivity when necessary
- print some additional information in case of nonconvergence