Force Neumann boundary conditions on regional grids with open boundaries.
The following namelist settings are required for runs on regional grids
with open boundary conditions:

  , restart_ext    = .true.
  , ew_boundary_type  = 'open'
  , ns_boundary_type  = 'open'
  , restore_ice     = .true.

In addition, two rows/columns of cells around the regional grid edge
(one halo, one physical) must be defined for the arrays to which the
state variables are restored, in ice_HaloRestore_init.   The changes 
in this revision force Neumann boundary conditions along those edges, 
but only to the values defined by the initial ice state.  Its up to 
the user to define these boundaries differently, if desired.

Note that grid_type can be either rectangular or displaced_pole.  
Perhaps we should add a regional option with its own subroutine?

Additional notes from the r499 log:

Add routines for gathering, scattering and (unformatted) reading and writing 
of the "extended" global grid, including the physical domain and ghost cells 
around the outer edges. These modifications allow exact restarts on regional 
grids with open boundary conditions, and they will also simplify restarts on 
the various tripole grids. No netcdf changes yet.

Contrary to what the cicedoc says, Neumann conditions are not available for 
open boundaries---the current infrastructure is not flexible enough to allow 
both local copies and messages along the same edge, which would be necessary 
for Neumann conditions in corners. Instead, restoring is required to enforce 
open boundary conditions (as in POP).