Boundary conditions for restoring.
Added subroutine to ice_restoring.F90 that explicitly sets restoring arrays.
NOTE This routine is highly dependent on the particular regional grid 
configuration and desired boundary conditions.  Currently it is set up 
for S. Stelma at U. Delaware and kludges the initial ice state to be the 
same as the restoring state.
Fixed indexing and conditionals so that all boundaries are set for single 
processor/block cases.