Bug fix for mushy thermodynamics in brackish seas:
To get the initial ice salinity when new ice forms, a fixed amount of 3 ppt 
was subtracted from the sea surface salinity (SSS), causing negative ice 
salinities when SSS < 3 ppt.  This fix keeps the same 3 ppt reduction in 
salinity for all SSS > 6 ppt, but it makes the initial sea ice salinity go 
to 0 continuously (and differentiably) as SSS approaches 0. We also added 
an arbitrary -0.1 C upper limit to the new ice temperature.